pyproject.toml
[project]
name = "djpress-publish-mastodon"
version = "1.2.0"
description = "A plugin for DJ Press that publishes new posts to Mastodon."
readme = "README.md"
authors = [{ name = "Stuart Maxwell", email = "stuart@amanzi.nz" }]
requires-python = ">=3.10"
dependencies = ["djpress>=0.20", "mastodon-py>=1.8.1"]

[dependency-groups]
dev = ["bumpver>=2024.1130"]

[project.scripts]
djpress-publish-mastodon = "djpress_publish_mastodon:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120

[tool.bumpver]
current_version = "1.2.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "👍 bump version {old_version} -> {new_version}"
commit = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = ['version = "{version}"']
"src/djpress_publish_mastodon/__init__.py" = ['^__version__ = "{version}"$']
